@@ -41,13 +41,12 @@
 		"multer-storage-cloudinary": "^4.0.0",
 		"node-gyp": "^8.2.0",
 		"prop-types": "^15.7.2",
-		"query-string": "^7.0.1",
-		"react": "^16.13.1",
+		"react": "^17.0.2",
 		"react-bootstrap": "next",
-		"react-dom": "^16.13.1",
+		"react-dom": "^17.0.2",
 		"react-is": "^17.0.2",
 		"react-router-dom": "^5.3.0",
-		"react-scripts": "3.4.1",
+		"react-scripts": "^4.0.3",
 		"react-toastify": "^8.0.3",
 		"remixicon": "^2.5.0",
 		"styled-components": "^5.3.1",
