@@ -1,183 +1,150 @@
-import React, { Component } from "react";
+import React, { useEffect, useState } from "react";
 import _ from "lodash";
-import { Link } from "react-router-dom";
 import { toast } from "react-toastify";
 import { getMyAddresses, deleteAddress } from "../../services/orderService";
-import Pagination from "../common/pagination";
 import { paginate } from "../../utils/paginate";
-import SearchBar from "../common/searchBar";
+
+import Pagination from "../common/pagination";
 import Table from "../common/table";
-import Like from "../common/like";
+import { TextInput } from "../common/inputs";
+import { Button } from "../common/buttons";
+import { Card1, CardBody1, CardHeader1 } from "../common/cards";
 
-class AddressBook extends Component {
-	state = {
-		addresses: [],
-		currentPage: 1,
-		pageSize: 4,
-		searchQuery: "",
-		selectedGenre: null,
-		sortColumn: { path: "title", order: "asc" },
-	};
+const AddressBook = () => {
+  const [addresses, setAddresses] = useState([]);
+  const [currentPage, setCurrentPage] = useState(1);
+  const [pageSize, setPageSize] = useState(4);
+  const [sortColumn, setSortColumn] = useState({ path: "title", order: "asc" });
+  const [data, setData] = useState({});
+  const [errors, setErrors] = useState({});
 
-	columns = [
-		{
-			path: "firstName",
-			label: "First Name",
-			// content: (address) => (
-			// 	<Link to={`/addresses/${address._id}`}>{address.product.title}</Link>
-			// ),
-		},
-		{ path: "lastName", label: "Last Name" },
-		{ path: "contactNo", label: "Contact No" },
-		{ path: "addLine1", label: "Address" },
-		{ path: "zipcode", label: "Zip Code" },
-		{
-			key: "like",
-			content: (address) => (
-				<Like liked={address.liked} onClick={() => this.props.onLike(address)} />
-			),
-		},
-		{
-			key: "delete",
-			content: (order) => (
-				<button
-					onClick={() => this.props.onDelete(order)}
-					className="btn btn-danger btn-sm"
-				>
-					Delete
-				</button>
-			),
-		},
-	];
+  const columns = [
+    { path: "firstName", label: "Name" },
+    { path: "contactNo", label: "Contact" },
+    { path: "city", label: "City" },
+    { path: "zipcode", label: "Zip" },
+  ];
 
-	async componentDidMount() {
-		// const { data } = await getGenres();
-		// const genres = [{ _id: "", name: "All Genres" }, ...data];
-		const addresses = await getMyAddresses();
-		this.setState({ addresses });
-	}
+  useEffect(() => {
+    populateAddresses();
+    toast("addressbook useEffect executed!");
+    addresses.filter((o) => {
+      console.log(o);
+    });
+  }, []);
 
-	handleDelete = async (address) => {
-		const originaladdresses = this.state.addresses;
-		const addresses = originaladdresses.filter((m) => m._id !== address._id);
-		this.setState({ addresses });
-		try {
-			await deleteAddress(address._id);
-		} catch (error) {
-			if (error.response && error.response.status === 404) {
-				toast.error("This address has already been deleted");
-				this.setState({ addresses: originaladdresses });
-			}
-		}
-	};
+  const populateAddresses = async () => {
+    const getaddresses = await getMyAddresses();
+    setAddresses(getaddresses);
+  };
 
-	handleLike = (movie) => {
-		console.log(this.state.addresses);
-		const addresses = [...this.state.addresses];
-		const index = addresses.indexOf(movie);
-		addresses[index] = { ...addresses[index] };
-		addresses[index].liked = !addresses[index].liked;
-		this.setState({ addresses });
-	};
+  const handleDelete = async (address) => {
+    const backup = addresses;
+    const filter = addresses.filter((m) => m._id !== address._id);
+    setAddresses(filter);
+    try {
+      await deleteAddress(address._id);
+    } catch (e) {
+      if (e.response && e.response.status === 404) {
+        toast.error("This address has already been deleted");
+        setAddresses(backup);
+      }
+    }
+  };
 
-	handlePageChange = (page) => {
-		this.setState({ currentPage: page });
-	};
+  const handlePageChange = (page) => {
+    setCurrentPage(page);
+  };
 
-	handleSearch = (query) => {
-		this.setState({
-			searchQuery: query,
-			selectedGenre: null,
-			currentPage: 1,
-		});
-	};
+  const handleSort = (sortColumn) => {
+    setSortColumn(sortColumn);
+  };
 
-	// handleGenreSelect = (genre) => {
-	// 	this.setState({
-	// 		selectedGenre: genre,
-	// 		searchQuery: "",
-	// 		currentPage: 1,
-	// 	});
-	// };
+  const getPagedData = () => {
+    const all = addresses;
 
-	handleSort = (sortColumn) => {
-		this.setState({ sortColumn });
-	};
+    const sorted = _.orderBy(all, [sortColumn.path], [sortColumn.order]);
 
-	getPagedData = () => {
-		const {
-			pageSize,
-			currentPage,
-			sortColumn,
-			// selectedGenre,
-			searchQuery,
-			addresses: allAddresses,
-		} = this.state;
+    const displayData = paginate(sorted, currentPage, pageSize);
 
-		let filtered = allAddresses;
-		if (searchQuery)
-			filtered = allAddresses.filter((o) =>
-				o.title.toLowerCase().startsWith(searchQuery.toLowerCase())
-			);
-		// else if (selectedGenre && selectedGenre._id)
-		// 	filtered = allAddresses.filter((o) => o.genre._id === selectedGenre._id);
+    return {
+      displayData,
+    };
+  };
 
-		const sorted = _.orderBy(filtered, [sortColumn.path], [sortColumn.order]);
+  const renderInput = (name, label, type = "text") => {
+    return (
+      <TextInput
+        type={type}
+        name={name}
+        value={data[name]}
+        label={label}
+        //   onChange={handleChange}
+        error={errors[name]}
+      />
+    );
+  };
 
-		const addresses = paginate(sorted, currentPage, pageSize);
+  const renderButton = (label) => {
+    return <Button className="mt-2">{label}</Button>;
+  };
 
-		return {
-			totalCount: filtered.length,
-			data: addresses,
-		};
-	};
+  const totalCount = addresses.length;
 
-	// constructor() {
-	// 	super();
-	// 	const user = authService.getCurrentUser();
-	// 	if (user && user.role === "admin") this.columns.push(this.deleteColumn);
-	// }
-	render() {
-		const { length: orderCount } = this.state.addresses;
-		const { pageSize, currentPage, sortColumn, searchQuery } = this.state;
-
-		if (orderCount === 0) return <p>There are no addresses in the database.</p>;
-
-		const { totalCount, data: addresses } = this.getPagedData();
-
-		return (
-			<div className="container-fluid mt-4">
-				{/* <div className="">
-					<ListGroup
-						items={this.state.genres}
-						selectedItem={this.state.selectedGenre}
-						onItemSelect={this.handleGenreSelect}
-					/>
-				</div> */}
-				<div>
-					<Link className="btn btn-primary mb-3" to="/addresses/new">
-						New Address
-					</Link>
-					<p>Showing {totalCount} addresses from the database</p>
-					<SearchBar value={searchQuery} onChange={this.handleSearch} />
-					<Table
-						columns={this.columns}
-						data={addresses}
-						sortColumn={sortColumn}
-						onSort={this.handleSort}
-						onLike={this.handleLike}
-						onDelete={this.handleDelete}
-					/>
-					<Pagination
-						itemsCount={totalCount}
-						pageSize={pageSize}
-						currentPage={currentPage}
-						onPageChange={this.handlePageChange}
-					/>
-				</div>
-			</div>
-		);
-	}
-}
+  const { displayData } = getPagedData();
+  return (
+    <div className="row">
+      <div className="col col-6">
+        {totalCount === 0 ? (
+          <Card1>
+            <CardHeader1 as="h6">
+              You haven't added any addresses, you're pretty anonymous!
+            </CardHeader1>
+          </Card1>
+        ) : (
+          <Card1>
+            <CardHeader1 as="h6">
+              {totalCount === 1 ? (
+                <>You've added an address!</>
+              ) : (
+                <>You've added {totalCount} addresses. Have Fun!!!</>
+              )}
+            </CardHeader1>
+            <CardBody1>
+              <Table
+                columns={columns}
+                data={displayData}
+                sortColumn={sortColumn}
+                onSort={handleSort}
+                onDelete={handleDelete}
+              />
+              <Pagination
+                itemsCount={totalCount}
+                pageSize={pageSize}
+                currentPage={currentPage}
+                onPageChange={handlePageChange}
+              />
+            </CardBody1>
+          </Card1>
+        )}
+      </div>
+      <div className="col col-6">
+        <Card1>
+          <CardHeader1 as="h5">Add an Address</CardHeader1>
+          <CardBody1>
+            <form>
+              {renderInput("productName", "Pla")}
+              {renderInput("user", "Customer")}
+              {renderInput("price", "Price")}
+              {renderInput("quantityVar", "Quantity")}
+              {renderInput("orderStatus", "We are currently")}
+              {renderButton("Add")}
+            </form>
+          </CardBody1>
+        </Card1>
+      </div>
+    </div>
+  );
+};
 
 export default AddressBook;
