@@ -2,8 +2,7 @@ import React from "react";
 import Joi from "joi-browser"; // a pretty sweet library for doing validation stuff in forms
 import Form from "../components/common/form";
 import authService from "../services/authService";
-import { CardHeader, ClassCard } from "../components/layout/card";
-import { Card } from "react-bootstrap";
+import { Card1, CardBody1, CardHeader1 } from "../components/common/cards";
 
 class SignupForm extends Form {
 	state = {
@@ -38,17 +37,17 @@ class SignupForm extends Form {
 	render() {
 		return (
 			<div className="justify-content-md-center row">
-				<ClassCard className="col col-lg-6">
-					<CardHeader as="h2">Register</CardHeader>
-					<Card.Body>
+				<Card1 className="col col-lg-6">
+					<CardHeader1 as="h2">Register</CardHeader1>
+					<CardBody1>
 						<form onSubmit={this.handleSubmit}>
 							{this.renderInput("name", "Name")}
 							{this.renderInput("username", "Username")}
 							{this.renderInput("password", "Password", "password")}
 							{this.renderButton("Register")}
 						</form>
-					</Card.Body>
-				</ClassCard>
+					</CardBody1>
+				</Card1>
 			</div>
 		);
 	}
